ipc-199: wip: ipc cli
