+++ ipc/cli/src/commands/checkpoint/topdown_executed.rs
@@ -7,9 +7,9 @@ use std::fmt::Debug;
 use async_trait::async_trait;
 use clap::Args;
 
-use crate::cli::commands::get_ipc_agent_url;
-use crate::cli::{CommandLineHandler, GlobalArguments};
+use crate::commands::get_ipc_agent_url;
 use crate::sdk::IpcAgentClient;
+use crate::{CommandLineHandler, GlobalArguments};
 
 /// The command to get the latest epoch executed for a top-down checkpoint
 pub(crate) struct LastTopDownExec;
