+++ ipc/cli/src/commands/crossmsg/fund.rs
@@ -6,9 +6,9 @@ use async_trait::async_trait;
 use clap::Args;
 use std::fmt::Debug;
 
-use crate::cli::commands::get_ipc_agent_url;
-use crate::cli::{CommandLineHandler, GlobalArguments};
+use crate::commands::get_ipc_agent_url;
 use crate::sdk::IpcAgentClient;
+use crate::{CommandLineHandler, GlobalArguments};
 
 /// The command to send funds to a subnet from parent
 pub(crate) struct Fund;
