+++ ipc/cli/src/commands/daemon.rs
@@ -12,10 +12,10 @@ use ipc_identity::Wallet;
 use tokio_graceful_shutdown::{IntoSubsystem, Toplevel};
 
 use crate::checkpoint::CheckpointSubsystem;
-use crate::cli::{CommandLineHandler, GlobalArguments};
 use crate::config::ReloadableConfig;
 use crate::server::jsonrpc::JsonRPCServer;
 use crate::server::{new_evm_keystore_from_config, new_fvm_wallet_from_config};
+use crate::{CommandLineHandler, GlobalArguments};
 
 /// The number of seconds to wait for a subsystem to start before returning an error.
 const SUBSYSTEM_WAIT_TIME_SECS: Duration = Duration::from_secs(10);
