+++ ipc/cli/src/commands/mod.rs
@@ -2,37 +2,37 @@
 // SPDX-License-Identifier: MIT
 //! This mod contains the different command line implementations.
 
-mod checkpoint;
-mod config;
-mod crossmsg;
-mod daemon;
+// mod checkpoint;
+// mod config;
+// mod crossmsg;
+// mod daemon;
 mod subnet;
 mod util;
-pub mod wallet;
+// pub mod wallet;
 
-use crate::cli::commands::checkpoint::CheckpointCommandsArgs;
-use crate::cli::commands::crossmsg::CrossMsgsCommandsArgs;
-use crate::cli::commands::daemon::{LaunchDaemon, LaunchDaemonArgs};
-use crate::cli::commands::util::UtilCommandsArgs;
-use crate::cli::{CommandLineHandler, GlobalArguments};
-use crate::server::{new_evm_keystore_from_path, new_keystore_from_path};
+// use crate::commands::checkpoint::CheckpointCommandsArgs;
+// use crate::commands::crossmsg::CrossMsgsCommandsArgs;
+// use crate::commands::daemon::{LaunchDaemon, LaunchDaemonArgs};
+use crate::commands::util::UtilCommandsArgs;
+// use crate::server::{new_evm_keystore_from_path, new_keystore_from_path};
+use crate::{CommandLineHandler, GlobalArguments};
 use anyhow::{Context, Result};
 
 use clap::{Command, CommandFactory, Parser, Subcommand};
 use clap_complete::{generate, Generator, Shell};
-use ipc_identity::{KeyStore, PersistentKeyStore};
+use fvm_shared::econ::TokenAmount;
 
 use std::fmt::Debug;
 use std::io;
+
 use subnet::SubnetCommandsArgs;
-use url::Url;
+// use crate::commands::config::ConfigCommandsArgs;
+// use crate::commands::wallet::WalletCommandsArgs;
 
-use crate::cli::commands::config::ConfigCommandsArgs;
-use crate::cli::commands::wallet::WalletCommandsArgs;
+// pub use subnet::*;
 
-pub use subnet::*;
-
-use super::default_repo_path;
+/// We only support up to 9 decimal digits for transaction
+const FIL_AMOUNT_NANO_DIGITS: u32 = 9;
 
 /// The collection of all subcommands to be called, see clap's documentation for usage. Internal
 /// to the current mode. Register a new command accordingly.
@@ -43,12 +43,12 @@ enum Commands {
     /// Note that, technically speaking, this just launches the ipc agent node and runs in the foreground
     /// and not in the background as what daemon processes are. Still, this struct contains `Daemon`
     /// due to the convention from `lotus` and the expected behavior from the filecoin user group.
-    Daemon(LaunchDaemonArgs),
-    Config(ConfigCommandsArgs),
+    // Daemon(LaunchDaemonArgs),
+    // Config(ConfigCommandsArgs),
     Subnet(SubnetCommandsArgs),
-    Wallet(WalletCommandsArgs),
-    CrossMsg(CrossMsgsCommandsArgs),
-    Checkpoint(CheckpointCommandsArgs),
+    // Wallet(WalletCommandsArgs),
+    // CrossMsg(CrossMsgsCommandsArgs),
+    // Checkpoint(CheckpointCommandsArgs),
     Util(UtilCommandsArgs),
 }
 
@@ -109,12 +109,12 @@ pub async fn cli() -> anyhow::Result<()> {
         let global = &args.global_params;
         if let Some(c) = &args.command {
             let r = match &c {
-                Commands::Daemon(args) => LaunchDaemon::handle(global, args).await,
-                Commands::Config(args) => args.handle(global).await,
+                // Commands::Daemon(args) => LaunchDaemon::handle(global, args).await,
+                // Commands::Config(args) => args.handle(global).await,
                 Commands::Subnet(args) => args.handle(global).await,
-                Commands::CrossMsg(args) => args.handle(global).await,
-                Commands::Wallet(args) => args.handle(global).await,
-                Commands::Checkpoint(args) => args.handle(global).await,
+                // Commands::CrossMsg(args) => args.handle(global).await,
+                // Commands::Wallet(args) => args.handle(global).await,
+                // Commands::Checkpoint(args) => args.handle(global).await,
                 Commands::Util(args) => args.handle(global).await,
             };
 
@@ -129,37 +129,24 @@ fn print_completions<G: Generator>(gen: G, cmd: &mut Command) {
     generate(gen, cmd, cmd.get_name().to_string(), &mut io::stdout());
 }
 
-pub(crate) fn get_ipc_agent_url(
-    ipc_agent_url: &Option<String>,
-    global: &GlobalArguments,
-) -> Result<Url> {
-    let url = match ipc_agent_url {
-        Some(url) => url.parse()?,
-        None => {
-            let config = global.config()?;
-            let addr = config.server.json_rpc_address.to_string();
-            // We are resolving back to our own ipc-agent node.
-            // Since it's our own node, we will use http since we
-            // should be in the same network.
-            format!("http://{addr:}/json_rpc").parse()?
-        }
-    };
-    Ok(url)
+pub(crate) fn get_ipc_provider(global: &GlobalArguments) -> Result<ipc_provider::IpcProvider> {
+    ipc_provider::IpcProvider::new_from_config(global.config_path())
 }
 
-pub(crate) fn get_fvm_store(path: Option<String>) -> Result<KeyStore> {
-    let path = match path {
-        Some(p) => p,
-        None => default_repo_path(),
-    };
-    new_keystore_from_path(&path)
+pub(crate) fn f64_to_token_amount(f: f64) -> anyhow::Result<TokenAmount> {
+    // no rounding, just the integer part
+    let nano = f64::trunc(f * (10u64.pow(FIL_AMOUNT_NANO_DIGITS) as f64));
+    Ok(TokenAmount::from_nano(nano as u128))
 }
 
-pub(crate) fn get_evm_keystore(
-    path: &Option<String>,
-) -> Result<PersistentKeyStore<ethers::types::Address>> {
-    match path {
-        Some(p) => new_evm_keystore_from_path(p),
-        None => new_evm_keystore_from_path(&default_repo_path()),
+#[cfg(test)]
+mod tests {
+    use crate::server::handlers::f64_to_token_amount;
+    use fvm_shared::econ::TokenAmount;
+
+    #[test]
+    fn test_amount() {
+        let amount = f64_to_token_amount(1000000.1f64).unwrap();
+        assert_eq!(amount, TokenAmount::from_nano(1000000100000000u128));
     }
 }
