+++ ipc/cli/src/commands/subnet/kill.rs
@@ -6,10 +6,10 @@ use async_trait::async_trait;
 use clap::Args;
 use std::fmt::Debug;
 
-use crate::cli::commands::get_ipc_agent_url;
-use crate::cli::{CommandLineHandler, GlobalArguments};
+use crate::commands::get_ipc_agent_url;
 use crate::sdk::IpcAgentClient;
 use crate::server::kill::KillSubnetParams;
+use crate::{CommandLineHandler, GlobalArguments};
 
 /// The command to kill an existing subnet.
 pub struct KillSubnet;
