+++ ipc/cli/src/commands/subnet/list_subnets.rs
@@ -10,10 +10,10 @@ use std::collections::HashMap;
 use std::fmt::Debug;
 use std::str::FromStr;
 
-use crate::cli::commands::get_ipc_agent_url;
-use crate::cli::{CommandLineHandler, GlobalArguments};
+use crate::commands::get_ipc_agent_url;
 use crate::config::json_rpc_methods;
 use crate::server::list_subnets::ListSubnetsParams;
+use crate::{CommandLineHandler, GlobalArguments};
 use ipc_provider::jsonrpc::{JsonRpcClient, JsonRpcClientImpl};
 use serde::Deserialize;
