+++ ipc/cli/src/lib.rs
@@ -6,11 +6,8 @@ use clap::Args;
 
 mod commands;
 
-use crate::config::Config;
 pub use commands::*;
-
-const DEFAULT_REPO_PATH: &str = ".ipc-agent";
-const DEFAULT_CONFIG_NAME: &str = "config.toml";
+use ipc_provider::config::Config;
 
 /// The trait that represents the abstraction of a command line handler. To implement a new command
 /// line operation, implement this trait and register it.
@@ -43,23 +40,13 @@ pub struct GlobalArguments {
 
 impl GlobalArguments {
     pub fn config_path(&self) -> String {
-        self.config_path.clone().unwrap_or_else(default_config_path)
+        self.config_path
+            .clone()
+            .unwrap_or_else(ipc_provider::default_config_path)
     }
 
     pub fn config(&self) -> Result<Config> {
         let config_path = self.config_path();
         Config::from_file(config_path)
     }
-}
-
-pub fn default_repo_path() -> String {
-    let home = match std::env::var("HOME") {
-        Ok(home) => home,
-        Err(_) => panic!("cannot get home"),
-    };
-    format!("{home:}/{:}", DEFAULT_REPO_PATH)
-}
-
-pub fn default_config_path() -> String {
-    format!("{}/{:}", default_repo_path(), DEFAULT_CONFIG_NAME)
 }
