+++ ipc/daemon/src/server/handlers/manager/mod.rs
@@ -29,33 +29,3 @@ pub mod subnet;
 pub mod topdown_executed;
 pub mod topdown_msgs;
 pub mod worker_addr;
-
-pub(crate) fn check_subnet(subnet: &Subnet) -> Result<()> {
-    match &subnet.config {
-        SubnetConfig::Fvm(config) => {
-            if config.auth_token.is_none() {
-                log::error!("subnet {:?} does not have auth token", subnet.id);
-                return Err(anyhow!("Internal server error"));
-            }
-        }
-        SubnetConfig::Fevm(_) => {
-            // TODO: add more checks later
-        }
-    }
-    Ok(())
-}
-
-pub(crate) fn parse_from(subnet: &Subnet, from: Option<String>) -> Result<Address> {
-    let addr = match from {
-        Some(addr) => Address::from_str(&addr)?,
-        None => {
-            if subnet.accounts().is_empty() {
-                log::error!("subnet does not have account defined, {:?}", subnet.id);
-                return Err(anyhow!("Internal server error"));
-            } else {
-                subnet.accounts()[0]
-            }
-        }
-    };
-    Ok(addr)
-}
